Дописать Gateway и доработать модули - проходит maven install

+++ commerce/order/src/main/java/ru/yandex/practicum/request/ProductReturnRequest.java
@@ -15,5 +15,5 @@ public class ProductReturnRequest {
 
     UUID orderId;
 
-    HashMap<UUID, Integer> products; // productId -> quantity;
+    HashMap<UUID, Long> products; // productId -> quantity;
 }
